# add neon simd version of find_authority_delimiter