Improves the performance of `find_authority_delimiter` on NEON architecture.

## Before

```
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations        GHz cycle/byte cycles/url instructions/byte instructions/cycle instructions/ns instructions/url     ns/url      speed  time/byte   time/url      url/s
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
BasicBench_AdaURL_href                  2184 ns         2177 ns       324169    4.95059      12.19   1.16889k           54.2804            4.45285         22.0442         5.20489k    236.111 396.448M/s   2.5224ns   241.87ns 4.13445M/s
BasicBench_AdaURL_aggregator_href       1325 ns         1320 ns       540499     5.9944     8.6825    832.556           40.0104            4.60817         27.6232         3.83656k    138.889 653.695M/s  1.52977ns  146.687ns 6.81721M/s
BasicBench_AdaURL_CanParse               941 ns          939 ns       742469    6.95086    7.04751    675.778           32.5736              4.622         32.1269         3.12344k    97.2222 919.306M/s  1.08778ns  104.306ns  9.5872M/s
BasicBench_whatwg                       4265 ns         4256 ns       165302    4.01608     19.387     1.859k           97.5875            5.03365         20.2156         9.35756k    462.889 202.782M/s  4.93141ns  472.868ns 2.11476M/s
BasicBench_CURL                        10933 ns        10922 ns        63206    3.55107    44.2341   4.24156k           218.441            4.93831         17.5363         20.9461k   1.19444k 79.0172M/s  12.6555ns  1.21352us  824.05k/s
```

## After

```
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
Benchmark                                  Time             CPU   Iterations        GHz cycle/byte cycles/url instructions/byte instructions/cycle instructions/ns instructions/url     ns/url      speed  time/byte   time/url      url/s
------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
BasicBench_AdaURL_href                  2189 ns         2187 ns       326079    4.79309    11.5689   1.10933k            54.343            4.69732         22.5146         5.21089k    231.444 394.551M/s  2.53453ns  243.033ns 4.11467M/s
BasicBench_AdaURL_aggregator_href       1317 ns         1316 ns       531616     5.9616    8.63499        828           39.9826             4.6303          27.604         3.83389k    138.889 655.728M/s  1.52502ns  146.233ns 6.83841M/s
BasicBench_AdaURL_CanParse               939 ns          939 ns       748015    7.04229    7.14021    684.667           32.5875            4.56394         32.1406         3.12478k    97.2222 919.365M/s  1.08771ns  104.299ns 9.58782M/s
BasicBench_whatwg                       4236 ns         4234 ns       165406    4.06455    19.6257   1.88189k           97.5689            4.97148         20.2069         9.35578k        463 203.843M/s  4.90574ns  470.406ns 2.12582M/s
BasicBench_CURL                        10873 ns        10870 ns        64574    3.54577     44.168   4.23522k           218.513            4.94732          17.542          20.953k   1.19444k 79.3933M/s  12.5955ns  1.20777us 827.972k/s
```