@lemire 2024-10-07
@anonrig I have pushed some fixes but I don't expect that this will make a measurable impact on the performance.