@anonrig 2024-10-10
@lemire I updated the PR description with the benchmark results. I'm not sure if it's worth merging. What do you think?